LPS-179353 Extract ScreenNavigationEntry methods from CommerceTermEntryQualifiersScreenNavigationCategory to CommerceTermEntryQualifiersScreenNavigationEntry, extends CommerceTermEntryQualifiersScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
